Alteração int para float

+++ Calculo_Média.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 
 int main(){
-    int nota1, nota2, nota3;
-    int media;
+    float nota1, nota2, nota3;
+    float media;
     
     printf("\n");
     printf("*** Programa de Cálculo de Média das Notas *** \n\n");
 
     printf("Digite a sua primeira nota: \n");
-    scanf("%d", &nota1);
+    scanf("%f", &nota1);
 
     printf("Digite a sua segunda nota: \n");
-    scanf("%d", &nota2);
+    scanf("%f", &nota2);
 
     printf("Digite a sua terceira nota: \n");
-    scanf("%d", &nota3);
+    scanf("%f", &nota3);
 
     media = (nota1 + nota2 + nota3) /3;
 
-    printf ("A Média é: %d", media);
+    printf ("A Média é: %.1f", media);
     
     
     return 0;
